+++ src/components/Comments.astro
@@ -1,4 +1,5 @@
 ---
+// src/components/Comments.astro
 export interface Props {
   owner: string;
   repo: string;
@@ -15,16 +16,15 @@ const { owner, repo, title } = Astro.props;
 </div>
 
 <script>
-  import { initializeComments } from '../utils/comments';
+  import { initializeComments } from '../utils/main';
 
   const renderApp = () => {
     const commentsContainer = document.getElementById('comments-container');
     if (!commentsContainer) return;
 
     const { owner, repo, title } = commentsContainer.dataset;
-    
     if (!owner || !repo || !title) {
-      console.error('Missing required data attributes');
+      console.error('Missing required data attributes for comments initialization.');
       return;
     }
 
@@ -36,77 +36,163 @@ const { owner, repo, title } = Astro.props;
 </script>
 
 <style>
-.user-info { 
-  display: flex; 
-  align-items: center; 
-  gap: 0.75rem; 
-  margin-bottom: 0.75rem; 
-}
-.user-info .user-avatar { 
-  border-radius: 50%; 
-}
-.user-info .user-name { 
-  font-weight: 500; 
-}
-.user-info .logout-btn { 
-  margin-left: auto; 
-  background: #d73a49; 
-  color: white; 
-  border: none; 
-  border-radius: 6px; 
-  padding: 0.4rem 0.8rem; 
-  font-size: 0.9rem; 
-  cursor: pointer; 
-}
-.user-info .logout-btn:hover { 
-  background: #cb2431; 
-}
-.comment-form-wrapper { 
-  display: flex; 
-  flex-direction: column; 
-  gap: 0.75rem; 
-  margin-top: 1.5rem; 
-}
-.user-avatar { 
-  border-radius: 50%; 
-}
-#comment-form { 
-  flex-grow: 1; 
-}
-#comment-form textarea {
-  width: 100%;
-  min-height: 80px;
-  padding: 0.5rem 0.75rem;
-  border: 1px solid #d1d5da;
-  border-radius: 6px;
-  font-size: 1rem;
-  line-height: 1.5;
-  resize: vertical;
-}
-.form-actions { 
-  margin-top: 0.5rem; 
-  text-align: right; 
-}
-.form-actions button {
-  background-color: #2ea44f;
-  color: white;
-  border: 1px solid rgba(27, 31, 35, 0.15);
-  border-radius: 6px;
-  padding: 0.5rem 1rem;
-  font-weight: 500;
-  cursor: pointer;
-}
-.form-actions button:hover { 
-  background-color: #2c974b; 
-}
-.form-actions button:disabled { 
-  background-color: #94d3a2; 
-  cursor: not-allowed; 
-}
-.login-prompt { 
-  text-align: center; 
-  margin-top: 1.5rem; 
-  padding-top: 1rem; 
-  border-top: 1px solid #e1e4e8; 
-}
+  .comment-loader {
+    text-align: center;
+    padding: 2rem 0;
+  }
+  
+  .error-box {
+    border: 1px solid #d73a49;
+    border-radius: 6px;
+    padding: 1rem;
+    background-color: #fef2f2;
+  }
+  
+  .error-message {
+    color: #cb2431;
+    font-family: monospace;
+  }
+
+  #comment-list {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+    margin-top: 1.5rem;
+  }
+
+  .comment {
+    border-left: 3px solid #d1d5da;
+    padding-left: 1.5rem;
+  }
+  
+  .comment-header {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    margin-bottom: 0.5rem;
+  }
+  
+  .user-avatar {
+    border-radius: 50%;
+  }
+
+  .comment-header .user-name {
+    font-weight: 500;
+    text-decoration: none;
+    color: inherit;
+  }
+  .comment-header .user-name:hover {
+    text-decoration: underline;
+  }
+
+  .comment-date {
+    margin-left: auto;
+    font-size: 0.875rem;
+    color: #586069;
+  }
+
+  .comment-body {
+    line-height: 1.6;
+  }
+  
+  .comment-form-wrapper {
+    margin-top: 2rem;
+    border-top: 1px solid #e1e4e8;
+    padding-top: 1.5rem;
+  }
+
+  .user-info {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    margin-bottom: 1rem;
+  }
+  
+  .user-info .user-name {
+    font-weight: 500;
+  }
+  
+  .user-info .logout-btn {
+    margin-left: auto;
+    background: #d73a49;
+    color: white;
+    border: none;
+    border-radius: 6px;
+    padding: 0.4rem 0.8rem;
+    font-size: 0.9rem;
+    cursor: pointer;
+  }
+  
+  .user-info .logout-btn:hover {
+    background: #cb2431;
+  }
+
+  #comment-form textarea {
+    width: 100%;
+    min-height: 100px;
+    padding: 0.75rem;
+    border: 1px solid #d1d5da;
+    border-radius: 6px;
+    font-size: 1rem;
+    line-height: 1.5;
+    resize: vertical;
+    box-sizing: border-box; /* 确保 padding 不会影响宽度 */
+  }
+  
+  .form-actions {
+    margin-top: 0.75rem;
+    text-align: right;
+  }
+
+  .form-actions button {
+    background-color: #2ea44f;
+    color: white;
+    border: 1px solid rgba(27, 31, 35, 0.15);
+    border-radius: 6px;
+    padding: 0.6rem 1.2rem;
+    font-weight: 500;
+    cursor: pointer;
+    transition: background-color 0.2s;
+  }
+  
+  .form-actions button:hover {
+    background-color: #2c974b;
+  }
+  
+  .form-actions button:disabled {
+    background-color: #94d3a2;
+    cursor: not-allowed;
+  }
+  
+  .login-prompt {
+    text-align: center;
+    margin-top: 1.5rem;
+    padding-top: 1.5rem;
+    border-top: 1px solid #e1e4e8;
+  }
+  
+  .login-prompt a {
+    font-weight: 500;
+    color: #0366d6;
+    text-decoration: none;
+  }
+  
+  .login-prompt a:hover {
+    text-decoration: underline;
+  }
+
+  .view-on-github {
+    text-align: right;
+    margin-top: 1rem;
+    font-size: 0.9rem;
+  }
+  
+  .view-on-github a {
+    color: #586069;
+    text-decoration: none;
+  }
+  
+  .view-on-github a:hover {
+    text-decoration: underline;
+  }
 </style>
